Renamed JackClientGRIS class to JackClient and fixed some compiler warnings.

+++ Source/Input.cpp
@@ -164,24 +164,24 @@ void Input::draw() const
     // conflict between glm, freeglut and juce's OpenGL implementations.
     //
     // This fix avoids calling glutSolidSphere().
-    static auto drawSphere = [](double const r) {
+    static auto drawSphere = [](float const r) {
         static auto constexpr LATS = 8;
         static auto constexpr LONGS = 8;
         for (int i{}; i <= LATS; ++i) {
             auto const lat0
-                = juce::MathConstants<double>::pi * (-0.5 + static_cast<double>(i - 1) / static_cast<double>(LATS));
+                = juce::MathConstants<float>::pi * (-0.5f + static_cast<float>(i - 1) / static_cast<float>(LATS));
             auto const z0 = std::sin(lat0);
             auto const zr0 = std::cos(lat0);
 
             auto const lat1
-                = juce::MathConstants<double>::pi * (-0.5 + static_cast<double>(i) / static_cast<double>(LATS));
+                = juce::MathConstants<float>::pi * (-0.5f + static_cast<float>(i) / static_cast<float>(LATS));
             auto const z1 = std::sin(lat1);
             auto const zr1 = std::cos(lat1);
 
             glBegin(GL_QUAD_STRIP);
             for (int j{}; j <= LONGS; ++j) {
                 auto const lng
-                    = 2.0 * juce::MathConstants<double>::pi * static_cast<double>(j - 1) / static_cast<double>(LONGS);
+                    = 2.0f * juce::MathConstants<float>::pi * static_cast<float>(j - 1) / static_cast<float>(LONGS);
                 auto const x = std::cos(lng);
                 auto const y = std::sin(lng);
